Multiple by scale for X size
The code as written divides the ```x``` value by the scale which results in a huge value, while the ```y``` value is multiplied by the scale.  Changing the x value to be multiplied produces better values and also prevents serious issues when loading tiles near the horizontal limits of the projection.